--- tests/rc/clk.config
@@ -1,1 +1,0 @@
-'tb_i.clk_gen'
