--- tests/rc/ctrl_io.config
@@ -1,4 +1,0 @@
-DigitalCtrlInput(name='go_vio', width=1, abspath='tb_i.go_vio', init_value=0x0)
-DigitalCtrlInput(name='rst_vio', width=1, abspath='tb_i.rst_vio', init_value=0x1)
-DigitalCtrlInput(name='v_in_vio', width=25, abspath='tb_i.v_in_vio', init_value=0x0)
-DigitalCtrlOutput(name='v_out_vio', width=25, abspath='tb_i.v_out_vio')
