--- tests/rc/probe.config
@@ -1,1 +1,0 @@
-['v_out_ila', 'tb_i.v_out', 25, 0]
