--- tests/rc/source.config
@@ -1,6 +1,0 @@
-VerilogSource(files="single_step_clk.sv")
-VerilogSource(files="tb.sv")
-VerilogSource(files="sim_ctrl.sv", fileset="sim")
-Define(name="DT_MSDSL", value="0.1e-6")
-Define(name="ANALOG_EXPONENT", value="-18")
-Define(name="SIMULATION_MODE_MSDSL", fileset="sim")
